+++ home-list/homelist-data/src/main/java/com/sendme/data/repository/bridge/FolderUpdateHandlerImpl.kt
@@ -1,13 +1,13 @@
 package com.sendme.data.repository.bridge
 
-import com.sendme.coredomain.navigation.bridge.FolderUpdateHandler
+import com.pingpad.coredomain.navigation.bridge.FolderUpdateHandler
 import com.sendme.data.db.FolderDao
 import javax.inject.Inject
 
 class FolderUpdateHandlerImpl @Inject constructor(
     private val folderDao: FolderDao
 ) : FolderUpdateHandler {
-    override suspend fun updateFolderWithLastNote(folderId: Int, lastNote: String, lastNoteDate: Long) {
+    override suspend fun updateFolderWithLastNote(folderId: Long, lastNote: String, lastNoteDate: Long) {
         folderDao.updateFolderLastNote(folderId = folderId, lastNote = lastNote, lastNoteDate = lastNoteDate)
     }
 }
